requirements.txt
ipykernel
numpy
pandas
pyyaml
openai